hf-backend/requirements.txt
gradio==4.31.3
transformers
torch
Pillow
sentence-transformers
faiss-cpu
huggingface-hub
opencv-python
requests
numpy
scikit-learn
